Move the preset related classes to a preset package
