+++ backend/src/main/java/com/benine/backend/ServerController.java
@@ -4,6 +4,8 @@ import com.benine.backend.camera.CameraController;
 import com.benine.backend.database.Database;
 import com.benine.backend.database.MySQLDatabase;
 import com.benine.backend.http.HTTPServer;
+import com.benine.backend.preset.Preset;
+import com.benine.backend.preset.PresetController;
 import com.benine.backend.video.StreamController;
 
 import java.io.File;
