+++ backend/src/main/java/com/benine/backend/http/PresetsHandler.java
@@ -1,7 +1,7 @@
 package com.benine.backend.http;
 
-import com.benine.backend.Preset;
-import com.benine.backend.PresetController;
+import com.benine.backend.preset.Preset;
+import com.benine.backend.preset.PresetController;
 import com.benine.backend.ServerController;
 import org.eclipse.jetty.server.Request;
 import org.json.simple.JSONArray;
