+++ backend/src/main/java/com/benine/backend/preset/Preset.java
@@ -1,4 +1,4 @@
-package com.benine.backend;
+package com.benine.backend.preset;
 
 import com.benine.backend.camera.Position;
 import org.json.simple.JSONArray;
