+++ backend/src/main/java/com/benine/backend/preset/PresetController.java
@@ -1,5 +1,6 @@
-package com.benine.backend;
+package com.benine.backend.preset;
 
+import com.benine.backend.ServerController;
 import com.benine.backend.database.Database;
 
 import java.sql.SQLException;
