+++ backend/src/test/java/com/benine/backend/http/AddTagHandlerTest.java
@@ -1,9 +1,7 @@
 package com.benine.backend.http;
 
-import com.benine.backend.PresetController;
+import com.benine.backend.preset.PresetController;
 import com.benine.backend.ServerController;
-import com.benine.backend.camera.CameraController;
-import com.sun.corba.se.spi.activation.Server;
 import org.eclipse.jetty.util.MultiMap;
 import org.junit.Before;
 import org.junit.Test;
