+++ backend/src/test/java/com/benine/backend/http/RecallPresetTest.java
@@ -1,6 +1,6 @@
 package com.benine.backend.http;
 
-import com.benine.backend.Preset;
+import com.benine.backend.preset.Preset;
 import com.benine.backend.camera.CameraConnectionException;
 import com.benine.backend.camera.Position;
 import com.benine.backend.camera.ipcameracontrol.IPCamera;
